requirements.txt
alabaster==0.7.9
arrow==0.8.0
astroid==1.4.8
awesome-slugify==1.6.5
Babel==1.3
backports.functools-lru-cache==1.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
binaryornot==0.4.0
blinker==1.3
chardet==2.3.0
click==6.6
configparser==3.5.0
cookiecutter==1.4.0
coreapi==2.0.8
coverage==4.2
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9
django-allauth==0.28.0
django-braces==1.9.0
django-coverage-plugin==1.3.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-environ==0.4.0
django-extensions==1.7.4
django-htmlmin==0.8.0
django-model-utils==2.6
django-redis==4.5.0
django-rest-swagger==2.0.7
django-tagulous==0.11.1
django-test-plus==1.0.16
django-toolbelt==0.0.1
django-tools==0.30.2
docutils==0.12
elasticsearch==2.4.0
enum34==1.1.6
extras==0.0.3
factory-boy==2.7.0
fake-factory==0.7.2
fixtures==2.0.0
flake8==3.0.4
Flask==0.11.1
Flask-Babel==0.11.1
Flask-Gravatar==0.4.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
future==0.15.2
gunicorn==19.6.0
html5lib==1.0b3
imagesize==0.7.1
importlib==1.0.3
ipaddress==1.0.17
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
isort==4.2.5
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.7.3
jinja2-time==0.2.0
lazy-object-proxy==1.2.2
linecache2==1.0.0
MarkupSafe==0.23
mccabe==0.5.2
mongoengine==0.10.6
neo4j-driver==1.0.2
oauthlib==2.0.0
openapi-codec==1.1.7
ordereddict==1.1
passlib==1.6.2
pathlib2==2.1.0
pbr==1.9.1
pdfminer==20140328
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
poyo==0.4.0
prompt-toolkit==1.0.8
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
pycodestyle==2.0.0
pycrypto==2.6.1
pyflakes==1.2.3
Pygments==2.1.3
pylint==1.6.4
pylint-common==0.2.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
pymongo==3.3.0
pyPdf==1.13
pyrsistent==0.11.13
pytest==3.0.3
pytest-django==3.0.0
python-dateutil==2.5.0
python-mimeparse==1.5.1
python-openid==2.2.5
pytz==2014.10
redis==2.10.5
regex==2016.10.22
requests==2.11.1
requests-oauthlib==0.7.0
simplegeneric==0.8.1
simplejson==3.6.5
six==1.9.0
snowballstemmer==1.2.1
speaklater==1.3
Sphinx==1.4.8
SQLAlchemy==1.0.14
sqlparse==0.2.0
static3==0.7.0
termcolor==1.1.0
testscenarios==0.5.0
testtools==2.0.0
traceback2==1.4.0
traitlets==4.3.1
Unidecode==0.4.19
unittest2==1.1.0
uritemplate==3.0.0
urllib3==1.18
wcwidth==0.1.7
Werkzeug==0.9.6
whichcraft==0.4.0
whitenoise==3.2.2
wrapt==1.10.8
WTForms==2.0.2
ipdb==0.10.1
pytest-django==3.0.0
pytest-sugar==0.7.1
